api-python/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# PDF Parsing
tabula-py>=2.9.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Brazilian Brokerage Notes
# Note: correpy may need to be installed from git if not on PyPI
# pip install git+https://github.com/victordalosto/correpy.git

# Data Processing
pandas>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# CORS
python-dotenv>=1.0.0
